+++ docs/source/tutorials/token_classification/run_ner.py
@@ -1,15 +1,16 @@
 from __future__ import absolute_import, division, print_function
-import json
-import os
-import torch
+
 import argparse
 import csv
+import json
 import logging
+import os
 import random
 import sys
-import torch.nn.functional as F
-from nltk import word_tokenize
+
 import numpy as np
+import torch
+import torch.nn.functional as F
 from pytorch_transformers import (WEIGHTS_NAME, AdamW, BertConfig,
                                   BertForTokenClassification, BertTokenizer,
                                   WarmupLinearSchedule)
@@ -18,208 +19,8 @@ from torch.utils.data import (DataLoader, RandomSampler, SequentialSampler,
                               TensorDataset)
 from torch.utils.data.distributed import DistributedSampler
 from tqdm import tqdm, trange
+
 from seqeval.metrics import classification_report
-
-############################################################## 
-
-def create_folds(sentences, k=10, path='folds/', seed=0): 
-    np.random.seed(seed) 
-    indicies = [i for i in range(len(sentences))] 
-    np.random.shuffle(indicies) 
-    partitions = [[sentences[index] for index in indicies[i::k]] for i in range(k)] 
-    
-    if os.path.exists(path): 
-        print("'%s' already exists, skipping..." % path) 
-    else: 
-        os.system('mkdir folds') 
-        for i in range(k): 
-            os.system('mkdir folds/fold%d' % i) 
-
-        def write_sentences_to_file(sentences, path): 
-            for sentence in sentences: 
-                for line in sentence: 
-                    path.write(line) 
-                path.write('\n')     
-
-        for i in range(k): 
-            train = open('folds/fold%d/train.txt' % i, "a") 
-            test = open('folds/fold%d/test.txt' % i, "a") 
-            for j in range(k): 
-                write_sentences_to_file(partitions[j], test if i == j else train) 
-            train.close() 
-            test.close() 
-        
-    indicies = [[index for index in indicies[i::k]] for i in range(k)] 
-    return indicies 
-
-
-def modified(given_words, sentence_tokens): 
-    for word, token in zip(given_words, sentence_tokens): 
-        if ''.join(word) != ''.join(token): 
-            return True  
-    return False 
-
-
-def get_probs(sentence, pipe, maps=None): 
-    def softmax(logit): 
-        return np.exp(logit) / np.sum(np.exp(logit)) 
-    
-    forward = pipe.forward(pipe.preprocess(sentence)) 
-    logits = forward['logits'][0].numpy() 
-    probs = np.array([softmax(logit) for logit in logits]) 
-    probs = probs[1:-1] 
-    
-    if not maps: 
-        return probs 
-    return merge_pros(probs, maps) 
-
-
-def get_pred_probs_and_labels(scores, tokens, given_token, given_label, weighted=False): 
-    i, j = 0, 0 
-    pred_probs, labels = [], [] 
-    for token, label in zip(given_token, given_label): 
-        i_new, j_new = i, j 
-        acc = 0 
-        
-        weights = []         
-        while acc != len(token): 
-            token_len = len(tokens[i_new][j_new:]) 
-            remain = len(token) - acc 
-            weights.append(min(remain, token_len)) 
-            if token_len > remain: 
-                acc += remain 
-                j_new += remain 
-            else: 
-                acc += token_len 
-                i_new += 1 
-                j_new = 0 
-        
-        if i != i_new: 
-            probs = np.average(scores[i:i_new], axis=0, weights=weights if weighted else None) 
-        else: 
-            probs = scores[i] 
-        i, j = i_new, j_new 
-        
-        pred_probs.append(probs) 
-        labels.append(label)
-        
-    return np.array(pred_probs), labels 
-
-
-def to_dict(nl): 
-    return {str(i): l for i, l in enumerate(nl)} 
-
-
-def read_npz(filepath): 
-    data = dict(np.load(filepath)) 
-    data = [data[str(i)] for i in range(len(data))] 
-    return data 
-
-############################################################## 
-
-class BertNer(BertForTokenClassification):
-
-    def forward(self, input_ids, token_type_ids=None, attention_mask=None, valid_ids=None):
-        sequence_output = self.bert(input_ids, token_type_ids, attention_mask, head_mask=None)[0]
-        batch_size,max_len,feat_dim = sequence_output.shape
-        valid_output = torch.zeros(batch_size,max_len,feat_dim,dtype=torch.float32,device='cuda' if torch.cuda.is_available() else 'cpu')
-        for i in range(batch_size):
-            jj = -1
-            for j in range(max_len):
-                    if valid_ids[i][j].item() == 1:
-                        jj += 1
-                        valid_output[i][jj] = sequence_output[i][j]
-        sequence_output = self.dropout(valid_output)
-        logits = self.classifier(sequence_output)
-        return logits
-
-class Ner:
-
-    def __init__(self,model_dir: str):
-        self.model , self.tokenizer, self.model_config = self.load_model(model_dir)
-        self.label_map = self.model_config["label_map"]
-        self.max_seq_length = self.model_config["max_seq_length"]
-        self.label_map = {int(k):v for k,v in self.label_map.items()}
-        self.device = "cuda" if torch.cuda.is_available() else "cpu"
-        self.model = self.model.to(self.device)
-        self.model.eval()
-
-    def load_model(self, model_dir: str, model_config: str = "model_config.json"):
-        model_config = os.path.join(model_dir,model_config)
-        model_config = json.load(open(model_config))
-        model = BertNer.from_pretrained(model_dir)
-        tokenizer = BertTokenizer.from_pretrained(model_dir, do_lower_case=model_config["do_lower"])
-        return model, tokenizer, model_config
-
-    def tokenize(self, text: str):
-        """ tokenize input"""
-        words = word_tokenize(text)
-        tokens = []
-        valid_positions = []
-        for i,word in enumerate(words):
-            token = self.tokenizer.tokenize(word)
-            tokens.extend(token)
-            for i in range(len(token)):
-                if i == 0:
-                    valid_positions.append(1)
-                else:
-                    valid_positions.append(0)
-        return tokens, valid_positions
-
-    def preprocess(self, text: str):
-        """ preprocess """
-        tokens, valid_positions = self.tokenize(text)
-        ## insert "[CLS]"
-        tokens.insert(0,"[CLS]")
-        valid_positions.insert(0,1)
-        ## insert "[SEP]"
-        tokens.append("[SEP]")
-        valid_positions.append(1)
-        segment_ids = []
-        for i in range(len(tokens)):
-            segment_ids.append(0)
-        input_ids = self.tokenizer.convert_tokens_to_ids(tokens)
-        input_mask = [1] * len(input_ids)
-        while len(input_ids) < self.max_seq_length:
-            input_ids.append(0)
-            input_mask.append(0)
-            segment_ids.append(0)
-            valid_positions.append(0)
-        return input_ids,input_mask,segment_ids,valid_positions
-
-    def predict(self, text: str):
-        input_ids,input_mask,segment_ids,valid_ids = self.preprocess(text)
-        input_ids = torch.tensor([input_ids],dtype=torch.long,device=self.device)
-        input_mask = torch.tensor([input_mask],dtype=torch.long,device=self.device)
-        segment_ids = torch.tensor([segment_ids],dtype=torch.long,device=self.device)
-        valid_ids = torch.tensor([valid_ids],dtype=torch.long,device=self.device)
-        with torch.no_grad():
-            logits = self.model(input_ids, segment_ids, input_mask,valid_ids)
-        logits = F.softmax(logits,dim=2)
-        pred_probs = logits.detach().cpu().numpy()[0] 
-        
-        logits_label = torch.argmax(logits,dim=2)
-        logits_label = logits_label.detach().cpu().numpy().tolist()[0]
-
-        logits_confidence = [values[label].item() for values,label in zip(logits[0],logits_label)]
-
-        logits = []
-        pos = 0
-        for index,mask in enumerate(valid_ids[0]):
-            if index == 0:
-                continue
-            if mask == 1:
-                logits.append((logits_label[index-pos],logits_confidence[index-pos]))
-            else:
-                pos += 1
-        logits.pop()
-
-        labels = [(self.label_map[label],confidence) for label,confidence in logits]
-        words = word_tokenize(text)
-        
-        pred_probs = pred_probs[1:len(words)+1] 
-        
-        return pred_probs, words  
 
 logging.basicConfig(format = '%(asctime)s - %(levelname)s - %(name)s -   %(message)s',
                     datefmt = '%m/%d/%Y %H:%M:%S',
